API/requirements.txt
Django==1.6.5
Fabric==1.9.0
Markdown==2.4.1
South==1.0
django-cors-headers==0.12
django-filter==0.7
djangorestframework==2.3.14
ecdsa==0.11
gondor==1.2.5
paramiko==1.14.0
#pycrypto==2.6.1 # already installed on WF
six==1.7.3
wsgiref==0.1.2
